+++ packages/web/src/app/my/register-club/[id]/page.tsx
@@ -1,219 +1,36 @@
 "use client";
 
-import React, { useMemo } from "react";
-
-import { RegistrationTypeEnum } from "@sparcs-clubs/interface/common/enum/registration.enum";
-import { ProfessorEnum } from "@sparcs-clubs/interface/common/enum/user.enum";
-import { useRouter } from "next/navigation";
-
-import styled from "styled-components";
+import React, { useEffect, useState } from "react";
 
 import NotFound from "@sparcs-clubs/web/app/not-found";
-import Button from "@sparcs-clubs/web/common/components/Button";
-import Card from "@sparcs-clubs/web/common/components/Card";
-import FlexWrapper from "@sparcs-clubs/web/common/components/FlexWrapper";
-import {
-  ListContainer,
-  ListItem,
-} from "@sparcs-clubs/web/common/components/ListItem";
-import PageHead from "@sparcs-clubs/web/common/components/PageHead";
-import Tag from "@sparcs-clubs/web/common/components/Tag";
-import Typography from "@sparcs-clubs/web/common/components/Typography";
-import { RegistrationTypeTagList } from "@sparcs-clubs/web/constants/tableTagList";
-import MyRegisterClubAcfTable from "@sparcs-clubs/web/features/my/register-club/components/MyRegisterClubAcfTable";
-import {
-  mockMyClubRegisterAcf,
-  mockMyClubRegisterDetail,
-} from "@sparcs-clubs/web/features/my/services/_mock/mockMyClubRegisterDetail";
 
-import { useGetProfileNow } from "@sparcs-clubs/web/hooks/getProfileNow";
-import { getTagDetail } from "@sparcs-clubs/web/utils/getTagDetail";
-
-const ButtonWrapper = styled.div`
-  display: flex;
-  gap: 16px;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-const TagWrapper = styled.div`
-  display: flex;
-  gap: 20px;
-  flex-direction: row;
-  justify-content: space-between;
-`;
+import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
+import LoginRequired from "@sparcs-clubs/web/common/frames/LoginRequired";
+import { useAuth } from "@sparcs-clubs/web/common/providers/AuthContext";
+import MyRegisterClubDetailFrame from "@sparcs-clubs/web/features/my/register-club/frames/MyRegisterClubDetailFrame";
 
 const MyRegisterClubDetail = () => {
-  const router = useRouter();
-  const profile = useGetProfileNow();
+  const { isLoggedIn, login, profile } = useAuth();
+  const [loading, setLoading] = useState(true);
 
-  const professorEnumToText = (profEnum: ProfessorEnum) => {
-    switch (profEnum) {
-      case ProfessorEnum.Assistant:
-        return "조교수";
-      case ProfessorEnum.Associate:
-        return "부교수";
-      case ProfessorEnum.Full:
-      default:
-        return "정교수";
+  useEffect(() => {
+    if (isLoggedIn !== undefined || profile !== undefined) {
+      setLoading(false);
     }
-  };
-  const data = useMemo(() => mockMyClubRegisterAcf.items, []);
+  }, [isLoggedIn, profile]);
 
-  if (profile === "executive") {
+  if (loading) {
+    return <AsyncBoundary isLoading={loading} isError />;
+  }
+
+  if (!isLoggedIn) {
+    return <LoginRequired login={login} />;
+  }
+
+  if (profile === "executive" || profile === undefined) {
     return <NotFound />;
   }
 
-  return (
-    <FlexWrapper direction="column" gap={60}>
-      <PageHead
-        items={[
-          { name: "마이페이지", path: "/my" },
-          {
-            name: "동아리 등록",
-            path: "/my/register-club",
-          },
-        ]}
-        title="동아리 등록"
-        enableLast
-      />
-      <Card outline gap={20}>
-        {/* TODO: Add Progress */}
-        <FlexWrapper direction="column" gap={16}>
-          <TagWrapper>
-            <Typography fw="MEDIUM" fs={16} lh={20}>
-              등록 구분
-            </Typography>
-            <Tag
-              color={
-                getTagDetail(
-                  mockMyClubRegisterDetail.registrationTypeEnum,
-                  RegistrationTypeTagList,
-                ).color
-              }
-            >
-              {
-                getTagDetail(
-                  mockMyClubRegisterDetail.registrationTypeEnum,
-                  RegistrationTypeTagList,
-                ).text
-              }
-            </Tag>
-          </TagWrapper>
-          <Typography fw="MEDIUM" fs={16} lh={20}>
-            기본 정보
-          </Typography>
-          <ListContainer>
-            <ListItem>동아리: {mockMyClubRegisterDetail.clubNameKr}</ListItem>
-            <ListItem>
-              대표자 이름: {mockMyClubRegisterDetail.studentName}
-            </ListItem>
-            <ListItem>
-              대표자 전화번호: {mockMyClubRegisterDetail.phoneNumber}
-            </ListItem>
-            <ListItem>
-              설립 연도: {mockMyClubRegisterDetail.foundedYear}
-            </ListItem>
-            <ListItem>
-              소속 분과: {mockMyClubRegisterDetail.clubDivision}
-            </ListItem>
-            <ListItem>
-              활동 분야 (국문): {mockMyClubRegisterDetail.activityFieldKr}
-            </ListItem>
-            <ListItem>
-              활동 분야 (영문): {mockMyClubRegisterDetail.activityFieldEn}
-            </ListItem>
-          </ListContainer>
-          <Typography fw="MEDIUM" fs={16} lh={20}>
-            지도교수 정보
-          </Typography>
-          <ListContainer>
-            <ListItem>성함: {mockMyClubRegisterDetail.professorName}</ListItem>
-            <ListItem>
-              직급:{" "}
-              {professorEnumToText(mockMyClubRegisterDetail.professorEnum)}
-            </ListItem>
-            <ListItem>
-              이메일: {mockMyClubRegisterDetail.professorEmail}
-            </ListItem>
-          </ListContainer>
-          <Typography fw="MEDIUM" fs={16} lh={20}>
-            동아리 정보
-          </Typography>
-          <ListContainer>
-            <ListItem>
-              분과 정합성: {mockMyClubRegisterDetail.divisionConsistency}
-            </ListItem>
-            <ListItem>
-              설립 목적: {mockMyClubRegisterDetail.foundationPurpose}
-            </ListItem>
-            <ListItem>
-              주요 활동 계획: {mockMyClubRegisterDetail.activityPlan}
-            </ListItem>
-            {mockMyClubRegisterDetail.registrationTypeEnum ===
-              RegistrationTypeEnum.Promotional && (
-              <ListItem>활동계획서</ListItem>
-            )}
-            {mockMyClubRegisterDetail.registrationTypeEnum ===
-              RegistrationTypeEnum.Promotional && (
-              <ListItem>동아리 회칙</ListItem>
-            )}
-
-            <ListItem>(선택) 외부 강사 지도 계획서</ListItem>
-          </ListContainer>
-          {/* TODO: Add File Preview */}
-          {mockMyClubRegisterDetail.registrationTypeEnum !==
-            RegistrationTypeEnum.Renewal && (
-            <FlexWrapper direction="column" gap={16}>
-              <Typography fw="MEDIUM" fs={16} lh={20}>
-                가등록 / 등록 취소 기간 활동 보고서 (총 {data.length}개)
-              </Typography>
-              <MyRegisterClubAcfTable
-                clubRegisterAcfList={mockMyClubRegisterAcf}
-              />
-            </FlexWrapper>
-          )}
-          <TagWrapper>
-            <Typography fw="MEDIUM" fs={16} lh={20}>
-              지도교수 승인
-            </Typography>
-            {mockMyClubRegisterDetail.professorConfirm ? (
-              <Tag color="GREEN">승인</Tag>
-            ) : (
-              <Tag color="GRAY">대기</Tag>
-            )}
-          </TagWrapper>
-        </FlexWrapper>
-      </Card>
-
-      <ButtonWrapper>
-        <Button
-          style={{ width: "max-content" }}
-          onClick={() => {
-            router.push("/my/register-club");
-          }}
-        >
-          목록으로 돌아가기
-        </Button>
-        {/* TODO: 집행부원은 다른 화면 보여야 하는지? */}
-        {profile !== "professor" ? (
-          <FlexWrapper direction="row" gap={10}>
-            <Button style={{ width: "max-content" }} onClick={() => {}}>
-              삭제
-            </Button>
-            <Button style={{ width: "max-content" }} onClick={() => {}}>
-              수정
-            </Button>
-          </FlexWrapper>
-        ) : (
-          <FlexWrapper direction="row" gap={10}>
-            <Button style={{ width: "max-content" }} onClick={() => {}}>
-              승인
-            </Button>
-          </FlexWrapper>
-        )}
-      </ButtonWrapper>
-    </FlexWrapper>
-  );
+  return <MyRegisterClubDetailFrame profile={profile} />;
 };
 export default MyRegisterClubDetail;
